Splitting command manipulation and processing functions into command.c. Lots of little cleanup and formatting fixes.
