+++ cmd-control-daemon/source/main.c
@@ -1,7 +1,6 @@
 #include <argp.h>
 #include <csp/csp.h>
 #include <csp/csp_interface.h>
-#include <dlfcn.h>
 #include <fcntl.h>
 #include <pthread.h>
 #include <stdlib.h>
@@ -11,11 +10,16 @@
 #include <unistd.h>
 
 #include "command-and-control/types.h"
+#include "cmd-control-daemon/daemon.h"
 
 #define PORT        10
 #define BUF_SIZE    250
 
-typedef void (*lib_func)();
+/*
+ * IMPORTANT: For review ignore everything before line #90
+ * We won't be using named pipes and everything before that is all csp code to
+ * set up a named pipe connection.
+ */
 
 pthread_t rx_thread, my_thread;
 int rx_channel, tx_channel;
@@ -28,6 +32,7 @@ csp_iface_t csp_if_fifo = {
     .mtu = BUF_SIZE,
 };
 
+
 int csp_fifo_tx(csp_iface_t *ifc, csp_packet_t *packet, uint32_t timeout) {
     /* Write packet to fifo */
     if (write(tx_channel, &packet->length, packet->length + sizeof(uint32_t) + sizeof(uint16_t)) < 0)
@@ -35,6 +40,7 @@ int csp_fifo_tx(csp_iface_t *ifc, csp_packet_t *packet, uint32_t timeout) {
     csp_buffer_free(packet);
     return CSP_ERR_NONE;
 }
+
 
 void * fifo_rx(void * parameters) {
     csp_packet_t *buf = csp_buffer_get(BUF_SIZE);
@@ -45,15 +51,6 @@ void * fifo_rx(void * parameters) {
     }
 
     return NULL;
-}
-
-
-
-int send_packet(csp_conn_t* conn, csp_packet_t* packet) {
-    printf("Sending: %s\r\n", packet->data);
-    if (!conn || !csp_send(conn, packet, 1000))
-        return -1;
-    return 0;
 }
 
 
@@ -90,9 +87,17 @@ int csp_init_things(int my_address){
     return 0;
 }
 
-//Where the magic happens
+//Where the magic happens - Bascially ignore everything above this line
 
-void send_response() {
+int send_packet(csp_conn_t* conn, csp_packet_t* packet) {
+    printf("Sending: %s\r\n", packet->data);
+    if (!conn || !csp_send(conn, packet, 1000))
+        return -1;
+    return 0;
+}
+
+
+void send_response(cnc_res_packet* response) { //TODO: Make this send an actual response packet
     int my_address = 1, client_address = 2;
     char *message = "Return MSG", *rx_channel_name, *tx_channel_name;
     csp_socket_t *sock;
@@ -115,87 +120,8 @@ void send_response() {
 }
 
 
-char* get_command(csp_socket_t* sock) {
-    csp_conn_t *conn;
-    csp_packet_t *packet;
-    char* command = NULL;
-
-    while (1) {
-        /* Process incoming packet */
-        conn = csp_accept(sock, 1000);
-        if (conn) {
-            packet = csp_read(conn, 0);
-            if (packet)
-                command = malloc(packet->length);
-                memcpy(command, packet->data, packet->length);
-                printf("Received Command: %s\r\n", command);
-            csp_buffer_free(packet);
-            csp_close(conn);
-            return command;
-       }
-    }
-}
-
-cnc_action parse_command(char * command) {
-    cnc_action action;
-    action = execute;
-    return action;
-}
-
-cnc_res_packet * run_command(cnc_cmd_packet * command){
-    int return_code;
-    cnc_res_packet* res_ptr = malloc(sizeof(cnc_res_packet));
-    void     *handle  = NULL;
-    lib_func  func    = NULL;
-    char * home_dir = "/home/vagrant/lib%s.so";
-
-    int len = strlen(home_dir) + strlen(command->cmd_name) - 1;
-
-    char * so_path = malloc(len);
-
-    snprintf(so_path, len, home_dir, command->cmd_name);
-
-    handle = dlopen(so_path, RTLD_NOW | RTLD_GLOBAL);
-
-    if (handle == NULL)
-    {
-        fprintf(stderr, "Unable to open lib: %s\n", dlerror());
-        return -1;
-    }
-
-    switch (command->action){
-        case execute:
-            printf("Running Command Execute\n");
-            func = dlsym(handle, "execute");
-            break;
-        case status:
-            printf("Running Command status\n");
-            func = dlsym(handle, "status");
-            break;
-        case version:
-            printf("Running Command version\n");
-            func = dlsym(handle, "version");
-            break;
-        case help:
-            printf("Running Command help\n");
-            func = dlsym(handle, "help");
-            break;
-        default:
-            printf ("Error the requested command doesn't exist\n");
-    }
-
-    if (func == NULL) {
-        fprintf(stderr, "Unable to get symbol\n");
-    }
-    printf("Calling fucntion with argc: %i, argv: %s\n", command->arg_count, command->args);
-    func(command->arg_count, command->args);
-    free(so_path);
-    return res_ptr;
-}
-
 int main(int argc, char **argv) {
     int my_address = 1;
-    char *message = "Testing CSP";
     csp_socket_t *sock;
     cnc_action action;
     cnc_cmd_packet* cmd_ptr = NULL;
@@ -212,7 +138,7 @@ int main(int argc, char **argv) {
         command = get_command(sock);
         cmd_ptr = parse(command);
         res_ptr = run_command(cmd_ptr);
-        send_response();
+        send_response(res_ptr);
         free(command);
     }
 
@@ -221,12 +147,10 @@ int main(int argc, char **argv) {
 
     //Free all the allocated things.
     int i;
-    for (i = 0; i < my_arguments.arg_count-3; i++){
-        free(my_arguments.args[i]);
+    for (i = 0; i < cmd_ptr->arg_count; i++){
+        free(cmd_ptr->args[i]);
     }
-    free(my_arguments.cmd_name);
-    free(result);
-
+    free(cmd_ptr->cmd_name);
 
     return 0;
 }
