+++ command-and-control/command-and-control/types.h
@@ -17,7 +17,7 @@ typedef struct arguments
     int arg_count;
     cnc_action action;
     char * cmd_name;
-    char *args[10];
+    char * args[10]; //TODO: Make a dynamic number of arguments
 } cnc_cmd_packet;
 
 
@@ -25,7 +25,7 @@ typedef struct
 {
     uint8_t return_code;
     double  execution_time;
-    char output[10]; //TODO: Make Dynamic
+    char    output[10]; //TODO: Make Dynamic
 } cnc_res_packet;
